Fixed potential crash with 'sw_delete' command
Using the 'sw_delete' command while the inventory screen was open would result in a crash if you flipped between inventory pages.
This was because the deleted item wasn't being purged properly from the inventory grid, so when the inventory tried to redraw the page it would include the deleted item, but since it didn't really exist the game would crash.